Refactor the ssh client code
We found out that during the softlayer-go attach/detach iSCSI volumes, the API will call ssh to execute some iSCSI scripts. This invocation process will cause some trouble in writing test cases for attach/detach persistent disk in bosh-softlayer-cpi project.

As a result, we refactor the shell script execution code process and move ssh client code to client package. After this code refactor, the shell script call from attach/detach API will be delegated to softlayer client, and the softlayer client will use the ssh client to finish the script execution.

After this change, now CPI project can use a fake ssh client which is inside the fake softlayer client in the test cases.
