+++ client/fakes/softlayer_client_fake.go
@@ -22,6 +22,8 @@ type FakeSoftLayerClient struct {
 
 	TemplatePath string
 
+	ExecShellCommandResult string
+
 	SoftLayerServices map[string]softlayer.Service
 
 	DoRawHttpRequestResponse       []byte
@@ -32,7 +34,7 @@ type FakeSoftLayerClient struct {
 	GenerateRequestBodyBuffer *bytes.Buffer
 	GenerateRequestBodyError  error
 
-	HasErrorsError, CheckForHttpResponseError error
+	HasErrorsError, CheckForHttpResponseError, ExecShellCommandError error
 }
 
 func NewFakeSoftLayerClient(username, apiKey string) *FakeSoftLayerClient {
@@ -200,6 +202,10 @@ func (fslc *FakeSoftLayerClient) HasErrors(body map[string]interface{}) error {
 
 func (fslc *FakeSoftLayerClient) CheckForHttpResponseErrors(data []byte) error {
 	return fslc.CheckForHttpResponseError
+}
+
+func (fslc *FakeSoftLayerClient) ExecShellCommand(username string, password string, ip string, command string) (string, error) {
+	return fslc.ExecShellCommandResult, fslc.ExecShellCommandError
 }
 
 //Private methods
